Rename stop_words.js to stop-words.js
Bring the file in line with the kebab-case convention used by every
other module under src/. Internal-only change — the exported function
and option keys (stopWords) are unchanged.
